Add files via upload
This study “the impact of taxation on corporate earnings and growth” with particular reference to guaranty trust bank plc yola branch is aimed at determining if taxation has an impact on corporate earnings and growth. The project work is divided into five (5) chapters, in chapter one (1), the researcher introduced the part that taxation played as a means of controlling the economy. the aim of this project is to verify the expected negative relationship between corporate taxation on the earning and long term economic growth in companies in the country using guaranty trust bank as a case study. this study leads to the findings of whether; (1) Tax has an impact on the earnings and growth on corporation using guaranty trust bank plc. As a case study (2) If Taxation has affected the financial and investment decision of corporation using guaranty trust bank plc as a study case. The chapter also presents the statement of the problems such as to what extend tax has contributed to the earnings and growth of corporate companies. Significance of the study was also discussed in this chapter, which is to verify the expected negative relationship between corporate taxation on the earning and long term economic growth in companies in the country using guaranty trust bank as a case study.
Chapter two (2) review the already analyze materials of some scholar to be relevant to the study, here, the meaning, objectives, importance of the study was given. Chapter three (3) discussed the research design, population of the study, sample size and sampling techniques used, method of data collection and analysis. The research study here is based on facts, information, data, oral interview and opinion gathered through informal questionnaire answered by the staff of guaranty trust bank plc yola branch. All the staff were use as the population of the study while a sample size of forty (40) was chosen, this figure was achieved through random sampling of total population.
Chapter four (4) discussed on the presentation, analysis of data generated through questionnaire, here distribution and return of questionnaire were shown in a table, the analysis shows that a total of twenty (20) questionnaire were distributed and the same total of twenty (20) were computed and returned.
Finally, chapter five (5) covers the summary, recommendation, limitation conclusion drawn from the study and suggestion of areas for further research was given here it was recommended that companies should prefer to finance their business through the use of debt and not tax deductible while on the other hand, dividends are subject to tax. It was also recommended that the change of account
 data should not be done indiscriminately as it has an adverse implication on financial decisions of corporation.